requirements.txt
openai>=1.30.0
pypdf>=4.2.0
PyYAML>=6.0.1
pytest>=8.2.0
streamlit>=1.44.0
